+++ FileSystem/Image.php
@@ -32,14 +32,14 @@ class Image extends File
 	/*** @var resource */
 	private $image = null;
 	/*** @var array */
-	private $exif = array();
+	private $exif = [];
 	/*** @var array */
-	static $imgFunctions = array(
+	static $imgFunctions = [
 			IMAGETYPE_GIF => 'imagecreatefromgif',
 			IMAGETYPE_JPEG => 'imagecreatefromjpeg',
 			IMAGETYPE_PNG => 'imagecreatefrompng',
 			IMAGETYPE_JPEG2000 => 'imagecreatefromjpeg'
-	);
+	];
 
 	/**
 	 * @param int $sections
@@ -49,7 +49,7 @@ class Image extends File
 	public function exif( $sections = 0, $array = true )
 	{
 		if ( function_exists( 'exif_read_data' ) && $this->_filename ) {
-			if ( in_array( strtolower( FileSystem::GetExt( $this->_filename ) ), array( 'jpg', 'jpeg', 'tiff' ) ) ) {
+			if ( in_array( strtolower( FileSystem::GetExt( $this->_filename ) ), [ 'jpg', 'jpeg', 'tiff' ] ) ) {
 				$this->exif = exif_read_data( $this->_filename, $sections, $array );
 			}
 			return $this->exif;
@@ -77,7 +77,7 @@ class Image extends File
 		$this->type = $imgType;
 		$currentImg = $this->createImage( $imgType );
 		if ( function_exists( 'imagecrop' ) ) {
-			$this->image = imagecrop( $currentImg, array( 'x' => $x, 'y' => $y, 'width' => $width, 'height' => $height ) );
+			$this->image = imagecrop( $currentImg, [ 'x' => $x, 'y' => $y, 'width' => $width, 'height' => $height ] );
 		}
 		else {
 			// imagecopy ( resource $dst_im , resource $src_im , int $dst_x , int $dst_y , int $src_x , int $src_y , int $src_w , int $src_h )